A number of places in the Makefile assume that the compiler is GCC and the linker is GNU ld. I've moved the GCC-specific CFLAGS to configure, and replaced $(LD) with $(CC) for linking because a non-GNU ld is not going to understand those options, but a compiler might (this also corrects a -Wl,... flag being passed to $(LD) when it's supposed to be for the compiler).
In the log_mapping structure, the index is an unsigned int but being initialised or compared to -1 in several cases. I've changed this to a size_t to be consistent with struct line's argc member in libirc since size_t is not guaranteed to be an unsigned int. I've added casts for the -1 initialisation.